Using unsigned for small indices

+++ src/test/test_code.cc
@@ -364,7 +364,7 @@ void add_rigidityE(const size_t nbt, const real* X, const real R1, real* Y)
     
     if ( nbt == DIM )
     {
-        for ( int d = 0; d < DIM; ++d )
+        for ( unsigned d = 0; d < DIM; ++d )
         {
             real x = 2 * X[d+DIM] - ( X[d+DIM*2] + X[d] );
             Y[d      ] += R1 * x;
@@ -383,7 +383,7 @@ void add_rigidityE(const size_t nbt, const real* X, const real R1, real* Y)
         for ( size_t i = DIM*2; i < end; ++i )
             Y[i] += R4 * (X[i-DIM]+X[i+DIM]) - R1 * (X[i-DIM*2]+X[i+DIM*2]) - R6 * X[i];
         
-        for ( size_t d = 0; d < DIM; ++d )
+        for ( unsigned d = 0; d < DIM; ++d )
         {
             Y[    d+DIM] -= R1 * (X[d+DIM]+X[d+DIM*3]) - R4 * (X[d+DIM*2]-X[d+DIM]) - R2 * X[d];
             Y[nbt+d    ] -= R1 * (E[d-DIM]+E[d-DIM*3]) - R4 * (E[d-DIM*2]-E[d-DIM]) - R2 * E[d];
@@ -409,7 +409,7 @@ void add_rigidityF(const size_t nbt, const real* X, const real R1, real* Y)
     real      * Z = Y + nbt + DIM;
     real const* E = X + nbt + DIM;
     #pragma ivdep
-    for ( size_t d = 0; d < DIM; ++d )
+    for ( unsigned d = 0; d < DIM; ++d )
     {
         Y[d    ] -= R1 * (X[d+DIM*2]+X[d]) - R2 * X[d+DIM];
         Y[d+DIM] -= R1 * (X[d+DIM]+X[d+DIM*3]) + R4 * (X[d+DIM]-X[d+DIM*2]) - R2 * X[d];
